Client_Farm 농장 데이터 저장 구조 변경
- 세이브 인덱스를 ENUM에서 INT를 사용하도록 변경

+++ NeoSports_Client/Assets/Scripts/Network/NetworkManager.cs
@@ -194,7 +194,7 @@ public class NetworkManager : Singleton<NetworkManager>
 		PACKET_REQ_RES_FARM packet = new PACKET_REQ_RES_FARM
 		{
 			header = headerPacket,
-			saveIndex = index,
+			saveIndex = (int)index,
 			saveData = data
 		};
